Fix the problem with client IP addresses being masked.
Using Docker CE 18.06 in it's default networking setup, swarm mode
appears incapable of delivering packets to containers with the source
address (eg of the original client) correctly set.

From reading through *many* issues about this on GitHub, it appears
that at times in the past it has worked.  However, the code and
functionality appears to suffer from many waves of advance/then
regression.  eg not really suitable for relying on

A commonly suggested workaround is to disable the userland proxy.

At least for Docker CE 18.06 on CentOS 7... this appears to have
no effect whatsoever.  Routing appears identical with it enabled
or disabled (as per config or dockerd command line), and the IP
address masking still happens.

One working approach is to instead use "host" mode networking.

That stops multiple containers being able to run in any host
(eg max of one per host binding to a given port), but at least it
works.  The client application in this instance is written in Go
and capable of scaling itself internally, so this probably isn't
much of a problem.